requirements.txt
numpy>=1.19.0
pandas>=1.2.0
matplotlib>=3.3.0
tensorflow>=2.5.0
plotly>=5.3.0
arch>=4.19.0
scikit-learn>=0.24.0
jupyter>=1.0.0
jupyterlab>=3.0.0
pandera>=0.8.0
mypy>=0.900
statsmodels>=0.12.0
scipy>=1.6.0
tqdm>=4.62.0
gridstatus>=0.16.0
pytz>=2022.1
requests>=2.27.1
beautifulsoup4>=4.10.0
python-dotenv>=0.20.0 